partly modified for commit

+++ runtime/megatron/data/indexed_dataset.py
@@ -95,9 +95,9 @@ dtypes = {
     3: np.int16,
     4: np.int32,
     5: np.int64,
-    6: np.float,
-    7: np.double,
-    8: np.uint16
+    6: np.float32,   # 或 np.float64，取决于原始数据格式定义
+    7: np.float64,   # 比 np.double 更直白
+    8: np.uint16,
 }
 
 
@@ -268,8 +268,8 @@ class IndexedDatasetBuilder(object):
         np.int16: 2,
         np.int32: 4,
         np.int64: 8,
-        np.float: 4,
-        np.double: 8
+        np.float32: 4,
+        np.float64: 8
     }
 
     def __init__(self, out_file, dtype=np.int32):
